--- a/lib/Cargo.toml
+++ b/lib/Cargo.toml
@@ -23,13 +23,14 @@ bench-suite = []
 
 [dependencies]
 byteorder = "^1.4.2"
-internment = "0.7.0"
 serde = { version = "^1.0.57", features = ["derive"] }
 serde_json = "^1.0.57"
 lazy_static = "^1.4.0"
 rmp-serde = "1.1.1"
 tempfile = "^3.2.0"
 uuid = { version = "^1.2.2", features = ["v1", "serde"] }
+once_cell = { version = "1.17", features = [ "parking_lot" ] }
+parking_lot = "0.12"
 
 # Rocksdb dependencies
 rocksdb = { version = "0.19.0", optional = true }