--- a/lib/Cargo.toml
+++ b/lib/Cargo.toml
@@ -22,6 +22,7 @@ test-suite = []
 bench-suite = []
 
 [dependencies]
+blake3 = "1.3"
 byteorder = "^1.4.2"
 serde = { version = "^1.0.57", features = ["derive"] }
 serde_json = "^1.0.57"