--- a/lib/Cargo.toml
+++ b/lib/Cargo.toml
@@ -18,11 +18,12 @@ path = "src/lib.rs"
 [features]
 default = []
 rocksdb-datastore = ["rocksdb", "bincode"]
+multi-slot = ["blake3"]
 test-suite = []
 bench-suite = []
 
 [dependencies]
-blake3 = "1.3"
+blake3 = { version = "1.3", optional = true }
 byteorder = "^1.4.2"
 serde = { version = "^1.0.57", features = ["derive"] }
 serde_json = "^1.0.57"