## What's this PR do
adds the new function `const_new_unchecked` to the Identifier struct, which enables the creation of an identifier with a constant string without any validation check.

## Motivation
Consider following codes:
```rust
mod identifiers {
    const PERSON: Identifier = unsafe { Identifier::const_new_unchecked("person") };
    const MOVIE: Identifier = unsafe { Identifier::const_new_unchecked("movie") };
}
```

This allows to create constant identifiers for encapsulation and abstraction.

## Changes
1. The addition of a new `MayStatic` enum type, which has two possible variants:
    * `Static`: A variant that holds a string reference to a constant `&'static str`.
    * `Interned`: A variant that holds an `Intern<String>` instance.
2. The modification of the existing `Identifier` struct to use the `MayStatic` enum instead of an `Intern<String>` instance to hold the value.