@ysimonson 2023-02-25
Support for const would be nice, but I'm not sure it's worth the cost of branching on all `Identifier` methods, since they're used so heavily. Could you outline the use case? Maybe there's an alternative.

As an aside, it is sort of silly that `Intern::new` isn't const, given it makes the internal value are `'static` anyways. I think a rigorous fix to support const would need to be applied there. But maybe there's a reason why it's not possible.