@lightsing 2023-02-27
> Support for const would be nice, but I'm not sure it's worth the cost of branching on all `Identifier` methods, since they're used so heavily. Could you outline the use case? Maybe there's an alternative.
> 
> As an aside, it is sort of silly that `Intern::new` isn't const, given it makes the internal value `'static` anyways. I think a rigorous fix to support const would need to be applied there. But maybe there's a reason why it's not possible.

You are right about the "branching" thing. Maybe we can find a better way.

The `Intern::new` can't be const since it need to allocate memory in the internal `HashSet`.

I'm new to graph db, so why the `Identifier` is usually build in runtime? As I'm trying out indradb, I pre-defined all possible `Identifier` in a mod via `once_cell::sync::Lazy` and use only their references.
**IF** most common way to use `Identifier` is pre-define, may be we can use a better way to hold the internal string.