@lightsing 2023-02-27
The crate internment is not designed for build from 'static, since it wants to manage all thing it holds and remove duplicate.

So, I remove the internment dep and impl a special version only holds 'static str for indradb.