@lightsing 2023-02-27
p.s. I suggest to change `lazy_static` to `once_cell`

https://github.com/indradb/indradb/blob/a774da9585a02eba07cd026ac130dda7bd565d89/lib/src/util.rs#L19-L21