@ysimonson 2023-02-27
> I'm new to graph db, so why the Identifier is usually build in runtime? As I'm trying out indradb, I pre-defined all possible Identifier in a mod via once_cell::sync::Lazy and use only their references.

I suppose runtime would be more common, since it's necessary when called via gRPC. That said, supporting both would be nice.

> p.s. I suggest to change lazy_static to once_cell

Makes sense, happy to take a PR with this change.