@ysimonson 2023-02-28
This seems to be a fair bit slower than internment. The microbenchmarks don't test for `Identifier` overhead, but nevertheless seeing an impact:

```
MASTER
test memory::bench_bulk_insert    ... bench:   8,422,419 ns/iter (+/- 1,077,473)
test memory::bench_create_edge    ... bench:          51 ns/iter (+/- 3)
test memory::bench_create_vertex  ... bench:         199 ns/iter (+/- 13)
test memory::bench_get_edge_count ... bench:          91 ns/iter (+/- 5)
test memory::bench_get_edges      ... bench:         227 ns/iter (+/- 11)
test memory::bench_get_vertices   ... bench:         209 ns/iter (+/- 10)

PR #282
test memory::bench_bulk_insert    ... bench:   9,451,130 ns/iter (+/- 1,035,730)
test memory::bench_create_edge    ... bench:          59 ns/iter (+/- 6)
test memory::bench_create_vertex  ... bench:         209 ns/iter (+/- 22)
test memory::bench_get_edge_count ... bench:         111 ns/iter (+/- 83)
test memory::bench_get_edges      ... bench:         254 ns/iter (+/- 52)
test memory::bench_get_vertices   ... bench:         235 ns/iter (+/- 37)
```

Not sure, but I wonder if the overhead is coming from doing full string comparisons rather than cheap pointer comparisons.