@lightsing 2023-02-28
I got a different result on wsl Ubuntu 22.04.2 LTS.

```
MASTER
test memory::bench_bulk_insert    ... bench:  17,438,960 ns/iter (+/- 18,438,881)
test memory::bench_create_edge    ... bench:          75 ns/iter (+/- 121)
test memory::bench_create_vertex  ... bench:         653 ns/iter (+/- 418)
test memory::bench_get_edge_count ... bench:          90 ns/iter (+/- 59)
test memory::bench_get_edges      ... bench:         228 ns/iter (+/- 223)
test memory::bench_get_vertices   ... bench:         226 ns/iter (+/- 242)

PR #282
test memory::bench_bulk_insert    ... bench:   6,139,715 ns/iter (+/- 779,448)
test memory::bench_create_edge    ... bench:          36 ns/iter (+/- 11)
test memory::bench_create_vertex  ... bench:         204 ns/iter (+/- 34)
test memory::bench_get_edge_count ... bench:          54 ns/iter (+/- 8)
test memory::bench_get_edges      ... bench:         119 ns/iter (+/- 20)
test memory::bench_get_vertices   ... bench:         109 ns/iter (+/- 11)
```