@lightsing 2023-02-28
```
performance change after multi-slot
test memory::bench_bulk_insert    ... bench:   5,761,460 ns/iter (+/- 522,578)
test memory::bench_create_edge    ... bench:          36 ns/iter (+/- 5)
test memory::bench_create_vertex  ... bench:         199 ns/iter (+/- 44)
test memory::bench_get_edge_count ... bench:          50 ns/iter (+/- 9)
test memory::bench_get_edges      ... bench:         129 ns/iter (+/- 17)
test memory::bench_get_vertices   ... bench:         107 ns/iter (+/- 15)
```