@ysimonson 2023-02-28
Interesting. I think this warrants some more direct benchmarking. Will do so in #284.