@lightsing 2023-03-03
There's no noticable performance change.
```
MASTER
test benches::benches::bench_ident_comparison      ... bench:           9 ns/iter (+/- 2)
test benches::benches::bench_ident_new             ... bench:       2,407 ns/iter (+/- 471)
test benches::benches::bench_ident_new_unchecked   ... bench:       1,905 ns/iter (+/- 328)
test benches::benches::bench_ident_renew           ... bench:         500 ns/iter (+/- 95)
test benches::benches::bench_ident_renew_unchecked ... bench:          63 ns/iter (+/- 13)

PR #282 mutli-slot disabled
test benches::benches::bench_ident_comparison      ... bench:          14 ns/iter (+/- 4)
test benches::benches::bench_ident_new             ... bench:       2,312 ns/iter (+/- 504)
test benches::benches::bench_ident_new_unchecked   ... bench:       1,746 ns/iter (+/- 456)
test benches::benches::bench_ident_renew           ... bench:         515 ns/iter (+/- 93)
test benches::benches::bench_ident_renew_unchecked ... bench:         104 ns/iter (+/- 14)
```
`multi-slot` is optimized for parallel `new`, but current benchmark code cannot bench this.