@ysimonson 2023-03-03
I'm still seeing a fairly noticeable performance degradation on a macbook pro / intel i5:

```
MASTER
test benches::benches::bench_ident_comparison      ... bench:          11 ns/iter (+/- 1)
test benches::benches::bench_ident_new             ... bench:       1,955 ns/iter (+/- 311)
test benches::benches::bench_ident_new_unchecked   ... bench:       1,427 ns/iter (+/- 193)
test benches::benches::bench_ident_renew           ... bench:         346 ns/iter (+/- 103)
test benches::benches::bench_ident_renew_unchecked ... bench:          92 ns/iter (+/- 15)

PR #282
test benches::benches::bench_ident_comparison      ... bench:          23 ns/iter (+/- 7)
test benches::benches::bench_ident_new             ... bench:       2,323 ns/iter (+/- 804)
test benches::benches::bench_ident_new_unchecked   ... bench:       1,665 ns/iter (+/- 994)
test benches::benches::bench_ident_renew           ... bench:         388 ns/iter (+/- 41)
test benches::benches::bench_ident_renew_unchecked ... bench:         156 ns/iter (+/- 74)
test memory::bench_bulk_insert                     ... bench:  10,255,757 ns/iter (+/- 4,132,138)
test memory::bench_create_edge                     ... bench:          59 ns/iter (+/- 10)
```